% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dist}
\alias{dist}
\title{distance data.}
\format{An object of class \code{matrix} with 35 rows and 35 columns.}
\usage{
dist
}
\description{
Contains distance matrix all region in Central Java
Data take from Central Java shapefile
source: bps.go.id
}
\keyword{datasets}

