% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{data example.}
\format{A data frame with twelve variables}
\usage{
example
}
\description{
Educational Data of Central Java in 2014
source: jateng.bps.go.id - educational data publication
}
\keyword{datasets}

