% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map}
\alias{map}
\title{map example.}
\format{A data frame with one variables: \code{Populasi}}
\usage{
map
}
\description{
Central Java shapefile
source: bps.go.id
}
\keyword{datasets}

