% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{population data.}
\format{A data frame with one variables: \code{Populasi}}
\usage{
population
}
\description{
Contains population data example
source: jateng.bps.go.id
}
\keyword{datasets}

