% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scalling.R
\name{scale}
\alias{scale}
\title{Data Scalling}
\usage{
scale(data, method = "zerotoone")
}
\arguments{
\item{data}{matrix data}

\item{method}{scalling technique use "z" for z-transform, "zerotoone" for zero to one scalling and "oneminuseone" minus one to one scalling}
}
\value{
scalled matrix data
}
\description{
Provide data scalling using z-transform, zero to one scalling and minus one to one scalling
}
\examples{
#load data
data <- example

#zero to one scalling
data <- scale(data)
data <- scale(data,method="zerotoone")

#z-transform
data <- scale(data,method="z")

#minus one to one scalling
data <- scale(data,method="oneminuseone")

}
\seealso{
\code{\link{fgwc}} for standard Fuzzy Geographically Weighted Clustering,
\code{\link{fgwc.gsa}} for optimize using Gravitational Search Algorithm,
\code{\link{spClustIndex}} for cluser validation,
\code{\link{visualize}} for cluster visualizatiion
}

