% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spClustIndex.R
\name{spClustIndex}
\alias{spClustIndex}
\title{Cluster Validity Index}
\usage{
spClustIndex(fgwc)
}
\arguments{
\item{fgwc}{result(object) from fgwc clustering}
}
\value{
validity indeks
}
\description{
This function used to validate the clustering result
}

