% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Cluster Visualization}
\usage{
visualize(fgwc)
}
\arguments{
\item{fgwc}{result(object) from fgwc clustering}
}
\value{
biPlot

radarPlot

clusterMap
}
\description{
This function visualize the clustering result
}

