% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spClustIndex.R
\name{spClustIndex}
\alias{spClustIndex}
\title{Cluster Validity Index}
\usage{
spClustIndex(fgwc)
}
\arguments{
\item{fgwc}{result(object) from fgwc clustering}
}
\value{
validity indeks
}
\description{
This function used to validate the clustering result
}
\examples{
#load data example
X <- example

#if using matrix distance
distance <- dist

#if using shapefile
#library(rgdal) for call readOGR
#distance <- readOGR(dsn = 'folder/.',"shapefile name")

#load population data
pop <- population

clust <- fgwc(X,pop,distance,K=2,m=1.5,beta=0.5)

#show cluster validation
spClustIndex(clust)


}
\seealso{
\code{\link{visualize}} for cluster visualizatiion
\code{\link{scale}} for data scalling
}

